+++ src/connect.rs
@@ -1,0 +1,5 @@
+use crate::Error;
+
+pub fn connect() -> Result<(), Error> {
+    todo!()
+}
