+++ src/scan.rs
@@ -1,0 +1,6 @@
+use crate::Error;
+use crate::api::ScanArgs;
+
+pub fn scan(args: ScanArgs) -> Result<(), Error> {
+    todo!()
+}
